requirements.txt
flask
openai
requests
python-dotenv
beautifulsoup4
google-api-python-client
